Move testing code to its own package
