+++ metricbeat/module/kubernetes/apiserver/apiserver_test.go
@@ -5,14 +5,14 @@ package apiserver
 import (
 	"testing"
 
-	"github.com/elastic/beats/metricbeat/helper/prometheus"
+	"github.com/elastic/beats/metricbeat/helper/prometheus/ptest"
 )
 
 const testFile = "_meta/test/metrics"
 
 func TestEventMapping(t *testing.T) {
-	prometheus.TestMetricSet(t, "kubernetes", "apiserver",
-		prometheus.TestCases{
+	ptest.TestMetricSet(t, "kubernetes", "apiserver",
+		ptest.TestCases{
 			{
 				MetricsFile:  "./_meta/test/metrics",
 				ExpectedFile: "./_meta/test/metrics.expected",
